+++ libXaw/src/AsciiSrc.c
@@ -1784,7 +1784,7 @@ GetDefaultPieceSize(Widget w, int offset, XrmValue *value)
     static XPointer pagesize;
 
     if (pagesize == NULL) {
-	pagesize = (XPointer)((long)_XawGetPageSize());
+	pagesize = (XPointer)((uintptr_t)_XawGetPageSize());
 	if (pagesize < (XPointer)BUFSIZ)
 	    pagesize = (XPointer)BUFSIZ;
     }
